Add Exception Middleware

+++ Application/Activities/Delete.cs
@@ -1,4 +1,5 @@
-﻿using Domain;
+﻿using Application.Core;
+using Domain;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using Persistence;
@@ -7,12 +8,12 @@ namespace Application.Activities;
 
 public class Delete
 {
-    public class Command : IRequest
+    public class Command : IRequest<Result<Unit>>
     {
         public Guid Id { get; set; }
     }
 
-    public class Handler : IRequestHandler<Command>
+    public class Handler : IRequestHandler<Command, Result<Unit>>
     {
         private readonly DataContext _context;
 
@@ -21,11 +22,17 @@ public class Delete
             _context = context;
         }
 
-        public async Task Handle(Command request, CancellationToken cancellationToken)
+        public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
         {
-             var activity = await _context.Activities.FirstOrDefaultAsync(x => x.Id == request.Id);
-             _context.Activities.Remove(activity);
-             await _context.SaveChangesAsync();
+            var activity = await _context.Activities.FirstOrDefaultAsync(x => x.Id == request.Id);
+            if (activity == null) return null;
+
+            _context.Activities.Remove(activity);
+            var result = await _context.SaveChangesAsync() > 1;
+
+            if (!result) return Result<Unit>.Failure("Failed to delete the activity.");
+
+            return Result<Unit>.Success(Unit.Value);
         }
     }
 }
